+++ lib/services/index.ts
@@ -2,11 +2,11 @@
  * Barrel export de servicios
  */
 export { lookupProfileByRut, getOrCreateProfile, getProfileByUserId, getProfileByEmail, linkProfileToUser, updateProfileDatosBase, saveTenantProfileData, getTenantProfileData, buildMergedAutofillData, computeTenantProfileStatus } from './profiles';
-export { createRegistration, createBulkRegistrations, listRegistrations, updateRegistrationStatus, bulkUpdateStatus, bulkDelete, getRegistrationFull, getRegistrationsByProfile, getRegistrationStats } from './registrations';
+export { createRegistration, createBulkRegistrations, listRegistrations, updateRegistrationStatus, bulkUpdateStatus, bulkDelete, getRegistrationFull, getRegistrationTenantId, getRegistrationsByProfile, getRegistrationStats } from './registrations';
 export { checkQuota, getQuotaRulesWithUsage, upsertQuotaRule, deleteQuotaRule } from './quotas';
 export { resolveZone, getZoneRules, upsertZoneRule, deleteZoneRule } from './zones';
-export { getTenantBySlug, getTenantById, listTenants, createTenant, updateTenant, createTenantAdmin, listTenantAdmins, listActiveTenants, deleteTenant } from './tenants';
-export { getActiveEvent, getEventById, getEventFull, listEventsByTenant, listAllEvents, createEvent, updateEvent, deactivateEvent, deleteEvent } from './events';
+export { getTenantBySlug, getTenantById, listTenants, createTenant, updateTenant, createTenantAdmin, listTenantAdmins, updateTenantAdmin, deleteTenantAdmin, listSuperAdmins, updateSuperAdmin, deleteSuperAdmin, listActiveTenants, deleteTenant } from './tenants';
+export { getActiveEvent, getEventById, getEventTenantId, getEventFull, listEventsByTenant, listAllEvents, createEvent, updateEvent, deactivateEvent, deleteEvent } from './events';
 export { listEventDays, getCurrentEventDay, createEventDay, createEventDaysBulk, updateEventDay, deleteEventDay, syncEventDays, getRegistrationDays, getEventDayCheckinStats } from './eventDays';
 export { getTeamMembers, addTeamMember, removeTeamMember, updateTeamMember, getTeamMembersForEvent } from './teams';
 export { sendApprovalEmail, sendRejectionEmail, sendBulkApprovalEmails, sendWelcomeEmail, sendInvitationEmail } from './email';
